+++ src/lib/api/leetcode.ts
@@ -3,11 +3,25 @@ import type {
   LeetcodeProblem,
   YamlProblem,
 } from "@/types/api/leetcode";
+import { markdownToBlocks } from "@tryfabric/martian";
+import { parseBlocks } from "../notion/utils";
+import { BlockObjectResponse } from "@notionhq/client/build/src/api-endpoints";
+import { v4 as uuid } from "uuid";
+import { Block } from "@/types/api/blocks";
 
 class Leetcode {
   private async readGitHubFile(path: string): Promise<string> {
+    const token = process.env.GITHUB_TOKEN;
+    if (!token) {
+      throw new Error("GITHUB_TOKEN is not set");
+    }
     const res = await fetch(
       `https://api.github.com/repos/ammar-ahmed22/lcgo/contents/${path}`,
+      {
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+      },
     );
     if (!res.ok) {
       throw new Error(`Failed to fetch ${path}`);
@@ -30,14 +44,29 @@ class Leetcode {
     );
 
     const leetcodeProblems: LeetcodeProblem[] = await Promise.all(
-      publishedProblems.map(async (entry) => {
-        const path = `${encodeURIComponent(entry[1].directory)}/docs.md`;
+      publishedProblems.map(async ([id, problem]) => {
+        const path = `${encodeURIComponent(problem.directory)}/docs.md`;
         const content = await this.readGitHubFile(path);
+        const blocks = markdownToBlocks(content);
+        const parsed = (
+          await parseBlocks(blocks as BlockObjectResponse[])
+        ).map((block) => {
+          return {
+            ...block,
+            id: block.id ?? uuid(),
+          };
+        });
+        // Find the first block that is not a heading
+        const description = parsed.find((block) => {
+          return block.type !== "heading";
+        });
         return {
-          id: entry[0],
-          difficulty: entry[1].difficulty,
-          name: entry[1].directory.split("-")[1].trim(),
-          content,
+          id,
+          difficulty: problem.difficulty,
+          name: problem.directory.split("-")[1].trim(),
+          raw: content,
+          blocks: parsed,
+          description: description as Block | undefined,
         };
       }),
     );
